stop mu tagging + change trunc. algos to use float --DC

+++ 3DMichel/TruncMean.cxx
@@ -3,8 +3,8 @@
 
 #include "TruncMean.h"
 
-void TruncMean::CalcTruncMean(const std::vector<double>& rr_v, const std::vector<double>& dq_v,
-			      std::vector<double>& dq_trunc_v)
+void TruncMean::CalcTruncMean(const std::vector<float>& rr_v, const std::vector<float>& dq_v,
+			      std::vector<float>& dq_trunc_v)
 {
 
   // how many points to sample 
@@ -18,7 +18,7 @@ void TruncMean::CalcTruncMean(const std::vector<double>& rr_v, const std::vector
   for (size_t n=0; n < dq_v.size(); n++) {
 
     // current residual range
-    double rr = rr_v.at(n);
+    float rr = rr_v.at(n);
 
     int nmin = n - Nneighbor;
     int nmax = n + Nneighbor;
@@ -27,11 +27,11 @@ void TruncMean::CalcTruncMean(const std::vector<double>& rr_v, const std::vector
     if (nmax > Nmax) nmax = Nmax;
 
     // vector for local dq values
-    std::vector<double> dq_local_v;
+    std::vector<float> dq_local_v;
 
     for (size_t i=nmin; i < nmax; i++) {
       
-      double dr = rr - rr_v[i];
+      float dr = rr - rr_v[i];
       if (dr < 0) dr *= -1;
 
       if (dr > _rad) continue;
@@ -46,10 +46,10 @@ void TruncMean::CalcTruncMean(const std::vector<double>& rr_v, const std::vector
     }
     
     // calculate median and rms
-    double median = Median(dq_local_v);
-    double rms    = RMS(dq_local_v);
+    float median = Median(dq_local_v);
+    float rms    = RMS(dq_local_v);
 
-    double truncated_dq = 0.;
+    float truncated_dq = 0.;
     int npts = 0;
     for (auto const& dq : dq_local_v) {
       if ( ( dq < (median+rms) ) && ( dq > (median-rms) ) ){
@@ -65,27 +65,27 @@ void TruncMean::CalcTruncMean(const std::vector<double>& rr_v, const std::vector
 }
 
 
-double TruncMean::Median(const std::vector<double>& v)
+float TruncMean::Median(const std::vector<float>& v)
 {
 
   if (v.size() == 1) return v[0];
   
-  std::vector<double> vcpy = v;
+  std::vector<float> vcpy = v;
 
   std::sort(vcpy.begin(), vcpy.end());
 
-  double median = vcpy[ vcpy.size() / 2 ];
+  float median = vcpy[ vcpy.size() / 2 ];
 
   return median;
 }
 
-double TruncMean::RMS(const std::vector<double>& v)
+float TruncMean::RMS(const std::vector<float>& v)
 {
 
-  double avg = 0.;
+  float avg = 0.;
   for (auto const& val : v) avg += val;
   avg /= v.size();
-  double rms = 0.;
+  float rms = 0.;
   for (auto const& val : v) rms += (val-avg)*(val-avg);
   rms = sqrt( rms / ( v.size() -  1 ) );
